requirements.txt
aiofiles==24.1.0
aiosmtplib==3.0.2
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.34.0
botocore==1.34.162
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
fastapi-mail==1.5.0
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic
python-multipart==0.0.6
PyYAML==6.0.2
redis==5.0.1
rsa==4.9.1
s3transfer==0.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.24.0
watchfiles==1.0.5
websockets==12.0
